move main insertion logic and job table update to database.insert module
