+++ src/database/insert.py
@@ -1,12 +1,17 @@
 import datetime
-from typing import Iterable
+import io
+import logging
+from typing import Iterable, Optional
 
 from geoalchemy2.shape import from_shape
 from pyproj import Transformer
 from shapely.geometry import Polygon, shape
 from shapely.ops import transform
+from sqlalchemy.exc import IntegrityError, NoResultFound
 from sqlalchemy.orm import Session
 
+from src import config
+from src.aws import s3
 from src.database.models import (
     AOI,
     Image,
@@ -18,6 +23,8 @@ from src.database.models import (
     SceneClassificationVector,
 )
 from src.models import DownloadResponse, Raster, Vector
+
+LOGGER = logging.getLogger(__name__)
 
 
 class Insert:
@@ -132,3 +139,66 @@ class Insert:
         self.session.bulk_save_objects(scls_vectors)
         self.session.commit()
         return scls_vectors
+
+
+class InsertJob:
+    def __init__(self, insert: Insert):
+        self.insert = insert
+
+    def insert_all(
+        self,
+        job_id: int,
+        download_response: DownloadResponse,
+        image: Raster,
+        pred_raster: Raster,
+        vectors: Iterable[Vector],
+    ) -> tuple[
+        Optional[Image], Optional[PredictionRaster], Optional[list[PredictionVector]]
+    ]:
+        unique_id = f"{download_response.bbox}/{download_response.image_id}"
+        image_url = s3.stream_to_s3(
+            io.BytesIO(download_response.content),
+            config.S3_BUCKET_NAME,
+            f"images/{unique_id}.tif",
+        )
+        try:
+            image_db = self.insert.insert_image(
+                download_response, image, image_url, job_id
+            )
+        except IntegrityError:
+            LOGGER.warning(f"Image {unique_id} already exists. Skipping")
+            return None, None, None
+
+        pred_raster_url = s3.stream_to_s3(
+            io.BytesIO(pred_raster.content),
+            config.S3_BUCKET_NAME,
+            f"predictions/{unique_id}.tif",
+        )
+        prediction_raster_db = self.insert.insert_prediction_raster(
+            pred_raster, image_db.id, pred_raster_url
+        )
+        prediction_vectors_db = self.insert.insert_prediction_vectors(
+            vectors, prediction_raster_db.id
+        )
+        return image_db, prediction_raster_db, prediction_vectors_db
+
+
+def set_init_job_status(db_session: Session, job_id: int, model_id: int):
+    model = db_session.query(Model).filter(Model.id == model_id).first()
+    if model is None:
+        update_job_status(db_session, job_id, JobStatus.FAILED)
+        raise NoResultFound("Model not found")
+    job = db_session.query(Job).filter(Job.id == job_id).first()
+
+    if job is None:
+        update_job_status(db_session, job_id, JobStatus.FAILED)
+        raise NoResultFound("Job not found")
+
+    else:
+        LOGGER.info(f"Updating job {job_id} to in progress")
+        update_job_status(db_session, job_id, JobStatus.IN_PROGRESS)
+
+
+def update_job_status(db_session: Session, job_id: int, status: JobStatus):
+    db_session.query(Job).filter(Job.id == job_id).update({"status": status})
+    db_session.commit()
